Draw circle based on 4 markers
Changed the code a bit (tentative) so it can now draw a relatively accurate circle if all four markers are identified.

Note that this accuracy will drop substantially if a frame is calculated where only three markers are seen.

(e.g. use frame0025 for four markers, and frame0028 for three markers to see the difference)