WIP: Scaling coordinate system to Optitrack
I think the scaling down works.

Plotting scale is still not correct, currently using hard-coded (300,300)

Astar also doesn't seem to work yet with the goal and starting points. Might be the way the path is calculated (it is calculated from goal to start instead of start to goal?) need fix